+++ tests/testthat/test-stacks.R
@@ -121,7 +121,7 @@ test_that("integration tests", {
   # If promises changes its internals, it can break this test on CRAN. Because
   # CRAN package releases are generally not synchronized (that is, promises and
   # shiny can't be updated at the same time, unless there is manual intervention
-  # from CRAN maintaineres), these specific test expectations make it impossible
+  # from CRAN maintainers), these specific test expectations make it impossible
   # to release a version of promises that will not break this test and cause
   # problems on CRAN.
   skip_on_cran()