+++ NEWS.md
@@ -33,6 +33,9 @@ shiny development
 
 * Closed tidyverse/dplyr#6154: Values from an `actionButton()` had S3 classes in the incorrect order.
 
+* Closed #2884 ('selectInput() turns NA into "NA"') by adding an option `keepNA` to `selectInput()` while maintaining previous behavior 
+with default value `keepNA = FALSE`.
+
 * Fixed a bug where updating an input value without a corresponding Input binding element did not trigger a JavaScript `shiny:inputchanged` event. Now, if no Input binding element is found, the `shiny:inputchanged` event is triggered on `window.document`. (#3584)
 
 * Restored the previous behavior of automatically guessing the `Content-Type` header for `downloadHandler` functions when no explicit `contentType` argument is supplied. (#3393)